+++ framework/modules/boot-starter/src/test/java/com/trc/test/notification/NotifyTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package group.idealworld.dew.notification;
+package com.trc.test.notification;
 
 
 import com.ecfront.dew.common.$;
